+++ dev/ext/lib/FileHandling/SP3/SP3Header.cpp
@@ -257,7 +257,7 @@ namespace gpstk
    try {
       SP3Stream& strm = dynamic_cast<SP3Stream&>(ffs);
       int i,k;
-      size_t j;
+      ssize_t j;
       string line;
       SP3SatID SVid;
       bool isVerA = (version == SP3a);
@@ -366,7 +366,7 @@ namespace gpstk
       //std::vector<std::string> comments; ///< vector of 4 comment lines
       for(j=0,i=19; i<=22; i++) {
          line = "/* ";
-         if(j < comments.size()) line += leftJustify(comments[j++],57);
+         if(j < (int)comments.size()) line += leftJustify(comments[j++],57);
          else line += string(57,'C');
          strm << line << endl;
          strm.lineNumber++;
